requirements.txt
adb-shell>=0.4.4
envsubst>=0.1.5
homekit[IP]>=0.19.0
jellyfin-apiclient-python>=1.9.2
pexpect>=4.9.0
phue>=1.1.0
Pillow>=10.1.0
py-sonic>=1.0.0
python-dotenv>=1.0.0
python-tado>=0.17.0
python-vlc>=3.0.0
pyvesync>=2.1.10
requests>=2.31.0
requests-cache>=1.1.1
retry-requests>=2.0.0
rgbxy>=0.5.0
streamdeck>=0.9.5
wheel>=0.42.0
yfinance>=0.2.36
